Refactor: add bird launch and scene input controllers, and integrate them into GameScene
