+++ include/GameScene.hpp
@@ -3,8 +3,9 @@
 
 #include "DynamicBackground.hpp"
 #include "LevelManager.hpp"
+#include "BirdLaunchController.hpp"
+#include "SceneInputController.hpp"
 #include "Scene.hpp"
-#include <vector>
 
 class GameScene : public Scene
 {
@@ -20,32 +21,18 @@ public:
         {
             m_DynamicBackground->SetSpeed(0.0f);
         }
+        m_SceneInputController = std::make_shared<SceneInputController>(m_DynamicBackground, m_LevelManager);
     }
 
     bool LoadLevel(const std::string &levelPath);
     void Update() override;
 
 private:
-    bool HandleBirdLaunchPhysics();
-    glm::vec2 GetMouseWorldPosition() const;
-    void ActivateBirdByIndex(size_t index);
-    void HandleBackgroundDrag();
-
     std::shared_ptr<LevelManager> m_LevelManager = std::make_shared<LevelManager>();
+    std::shared_ptr<BirdLaunchController> m_BirdLaunchController = std::make_shared<BirdLaunchController>();
+    std::shared_ptr<SceneInputController> m_SceneInputController = nullptr;
     std::shared_ptr<DynamicBackground> m_DynamicBackground = nullptr;
-    bool m_IsHoldingBackground = false;
-    bool m_IsDraggingBackground = false;
-    bool m_IsHoldingBird = false;
-    bool m_HasLaunchedBird = false;
-    float m_WorldOffsetX = 0.0f;
     float m_ZoomScrollAccumulator = 0.0f;
-    glm::vec2 m_BirdAnchorPosition{0.0f, 0.0f};
-    glm::vec2 m_BirdVelocity{0.0f, 0.0f};
-    glm::vec2 m_DragStartMousePos{0.0f, 0.0f};
-    glm::vec2 m_LastMousePos{0.0f, 0.0f};
-    std::shared_ptr<Character> m_ActiveBird = nullptr;
-    std::vector<std::shared_ptr<Character>> m_BirdQueue;
-    size_t m_CurrentBirdIndex = 0;
 };
 
 #endif // GAME_SCENE_HPP
